Implement whitelist, blacklist and unlimited access to faitout
Using the configuration file, one can now restrict the access to faitout to a
list of defined IP, or block a list of IP from ever getting a connection or
allow a list of IP to get as many active connection as they ask.

This allows to deploy faitout and use it without making it world accessible (via
the whitelist), block some spammers (via the blacklist) or integrate it with
some CI services like jenkins which may require multiple connections in short
period of times.